# api/swagger/openapi.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "title": "NATS Horizon API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "version": "1.0"
  },
  "basePath": "/api",
  "paths": {
    "/account/info": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "health"
        ],
        "summary": "Get JetStream account information",
        "responses": {
          "200": {
            "description": "Account information",
            "schema": {
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "description": "Returns all configured alert definitions",
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "List alerts",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Alert"
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a new alert configuration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "Create an alert",
        "parameters": [
          {
            "description": "Alert configuration",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/alerts/triggers": {
      "get": {
        "description": "Returns triggered alert instances, optionally filtered",
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "List alert triggers",
        "parameters": [
          {
            "type": "string",
            "description": "Filter by alert ID",
            "name": "alert_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Filter by acked state (true/false)",
            "name": "acked",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AlertTrigger"
              }
            }
          }
        }
      }
    },
    "/alerts/triggers/{id}/ack": {
      "post": {
        "description": "Acknowledges one or more triggers for the given alert ID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "Acknowledge an alert trigger",
        "parameters": [
          {
            "type": "string",
            "description": "Alert ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Acknowledging user",
            "name": "request",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          }
        }
      }
    },
    "/alerts/{id}": {
      "get": {
        "description": "Returns a single alert configuration by ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "Get an alert",
        "parameters": [
          {
            "type": "string",
            "description": "Alert ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "description": "Updates an existing alert configuration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "Update an alert",
        "parameters": [
          {
            "type": "string",
            "description": "Alert ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Alert configuration",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Deletes an alert configuration by ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "alerts"
        ],
        "summary": "Delete an alert",
        "parameters": [
          {
            "type": "string",
            "description": "Alert ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/cluster/info": {
      "get": {
        "description": "Returns JetStream cluster topology and server information",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "summary": "Get cluster information",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClusterInfoResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/cluster/streams/{name}/replicas": {
      "get": {
        "description": "Returns replication, mirror, source, and cluster placement info for a stream",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cluster"
        ],
        "summary": "Get stream replicas",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClusterStreamReplicaResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/connections": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "health"
        ],
        "summary": "Get NATS connections",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ConnectionsResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/connections/{id}": {
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "health"
        ],
        "summary": "Terminate a connection",
        "parameters": [
          {
            "type": "string",
            "description": "Connection ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/consumers": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "consumers"
        ],
        "summary": "List all consumers across all streams",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ConsumerResponse"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/consumers/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "consumers"
        ],
        "summary": "Get consumer by name across all streams",
        "parameters": [
          {
            "type": "string",
            "description": "Consumer name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ConsumerResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/core/monitor": {
      "get": {
        "description": "Opens a Server-Sent Events stream of traffic for the given subjects",
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "core-nats"
        ],
        "summary": "Monitor subject traffic (SSE stream)",
        "parameters": [
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "csv",
            "description": "Subjects to monitor (repeatable)",
            "name": "subjects",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/core/publish": {
      "post": {
        "description": "Publishes a message to a NATS subject (non-JetStream)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "core-nats"
        ],
        "summary": "Publish a Core NATS message",
        "parameters": [
          {
            "description": "Message to publish",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PublishMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PublishMessageResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/core/request": {
      "post": {
        "description": "Publishes a request message and waits for a reply (request/reply pattern)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "core-nats"
        ],
        "summary": "Send a Core NATS request",
        "parameters": [
          {
            "description": "Request message",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MessageResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "408": {
            "description": "Request Timeout",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "Service Unavailable",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/core/subscribe": {
      "get": {
        "description": "Opens a Server-Sent Events stream of messages published to the given subject",
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "core-nats"
        ],
        "summary": "Subscribe to a subject (SSE stream)",
        "parameters": [
          {
            "type": "string",
            "description": "NATS subject to subscribe to",
            "name": "subject",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/dashboard/stats": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "health"
        ],
        "summary": "Get dashboard statistics",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DashboardStatsResponse"
            }
          },
          "503": {
            "description": "Service Unavailable",
            "schema": {
              "$ref": "#/definitions/DashboardStatsResponse"
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "description": "Opens a Server-Sent Events stream for real-time stream/consumer/dashboard updates",
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "events"
        ],
        "summary": "Server-Sent Events stream",
        "parameters": [
          {
            "type": "string",
            "default": "all",
            "description": "Event channel (streams, consumers, dashboard, all)",
            "name": "channel",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/export/streams": {
      "get": {
        "description": "Exports a summary of all streams as JSON",
        "produces": [
          "application/json"
        ],
        "tags": [
          "export"
        ],
        "summary": "Export all streams",
        "responses": {
          "200": {
            "description": "Exported streams summary",
            "schema": {
              "type": "file"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/export/streams/{name}": {
      "get": {
        "description": "Exports stream data in the requested format (json, csv, txt)",
        "produces": [
          "application/json"
        ],
        "tags": [
          "export"
        ],
        "summary": "Export a stream",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "json",
            "description": "Export format (json, csv, txt)",
            "name": "format",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Include messages in the export",
            "name": "include_messages",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Exported stream data",
            "schema": {
              "type": "file"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/history/report": {
      "get": {
        "description": "Returns a summary report of all streams with latest metric values",
        "produces": [
          "application/json"
        ],
        "tags": [
          "history"
        ],
        "summary": "Get history report",
        "parameters": [
          {
            "type": "string",
            "default": "24h",
            "description": "Time window",
            "name": "duration",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "history report",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Alert": {
      "type": "object",
      "properties": {
        "channels": {
          "description": "Notification channels: \"email\", \"webhook\", \"slack\"",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "condition": {
          "$ref": "#/definitions/AlertCondition"
        },
        "cooldown": {
          "$ref": "#/definitions/Duration"
        },
        "created_at": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "id": {
          "type": "string"
        },
        "last_trigger": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "severity": {
          "$ref": "#/definitions/AlertSeverity"
        },
        "trigger_count": {
          "type": "integer"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  }
}